requirements.txt
azure-functions
scrapy
twisted